Refactor: Split memories widget into sepearate files
