+++ lib/ui/home/header_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
 import 'package:logging/logging.dart';
-import 'package:photos/ui/home/memories_widget.dart';
+import 'package:photos/ui/home/memories/memories_widget.dart';
 import 'package:photos/ui/home/status_bar_widget.dart';
 
 class HeaderWidget extends StatelessWidget {
