+++ lib/ui/home/memories/memories_widget.dart
@@ -1,0 +1,67 @@
+import 'package:flutter/material.dart';
+import 'package:photos/models/memory.dart';
+import 'package:photos/services/memories_service.dart';
+import "package:photos/ui/home/memories/memory_cover_widget.dart";
+
+class MemoriesWidget extends StatelessWidget {
+  const MemoriesWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return FutureBuilder<List<Memory>>(
+      future: MemoriesService.instance.getMemories(),
+      builder: (context, snapshot) {
+        if (snapshot.hasError || !snapshot.hasData || snapshot.data!.isEmpty) {
+          return const SizedBox.shrink();
+        } else {
+          return Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              _buildMemories(snapshot.data!),
+              const Divider(),
+            ],
+          );
+        }
+      },
+    );
+  }
+
+  Widget _buildMemories(List<Memory> memories) {
+    final collatedMemories = _collateMemories(memories);
+    final List<Widget> memoryWidgets = [];
+    for (final memories in collatedMemories) {
+      memoryWidgets.add(MemoryCovertWidget(memories: memories));
+    }
+    return SingleChildScrollView(
+      scrollDirection: Axis.horizontal,
+      child: Row(children: memoryWidgets),
+    );
+  }
+
+  List<List<Memory>> _collateMemories(List<Memory> memories) {
+    final List<Memory> yearlyMemories = [];
+    final List<List<Memory>> collatedMemories = [];
+    for (int index = 0; index < memories.length; index++) {
+      if (index > 0 &&
+          !_areMemoriesFromSameYear(memories[index - 1], memories[index])) {
+        final List<Memory> collatedYearlyMemories = [];
+        collatedYearlyMemories.addAll(yearlyMemories);
+        collatedMemories.add(collatedYearlyMemories);
+        yearlyMemories.clear();
+      }
+      yearlyMemories.add(memories[index]);
+    }
+    if (yearlyMemories.isNotEmpty) {
+      collatedMemories.add(yearlyMemories);
+    }
+    return collatedMemories.reversed.toList();
+  }
+
+  bool _areMemoriesFromSameYear(Memory first, Memory second) {
+    final firstDate =
+        DateTime.fromMicrosecondsSinceEpoch(first.file.creationTime!);
+    final secondDate =
+        DateTime.fromMicrosecondsSinceEpoch(second.file.creationTime!);
+    return firstDate.year == secondDate.year;
+  }
+}
